refector: replacing swiper for new carousel component
