+++ src/components/VideoCarousel/index.jsx
@@ -1,13 +1,12 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import useEmblaCarousel from 'embla-carousel-react'
 import styles from './VideoCarousel.module.css'
 import Videothumb from '../Videothumb'
+import EmblaCarousel from '../EmblaCarousel'
 
 const VideoCarousel = ({ playlist, hideLabel }) => {
   const { t } = useTranslation();
-  const [emblaRef] = useEmblaCarousel({ align: 'start' });
 
   return (
     <div>
@@ -23,23 +22,21 @@ const VideoCarousel = ({ playlist, hideLabel }) => {
         </Link>
       </div>
 
-      <div className={styles.emblaViewport} ref={emblaRef}>
-        <div className={styles.emblaContainer}>
-          {
-            playlist.videos.map((video) =>
-              <div className={styles.emblaSlide}>
-                <Videothumb
-                  key={video.videoId}
-                  videoName={video.name}
-                  videoId={video.videoId}
-                  videoLink={video.link}
-                  hideLabel={hideLabel}
-                />
-              </div>
-            )
-          }
-        </div>
-      </div>
+      <EmblaCarousel>
+        {
+          playlist.videos.map((video) =>
+            <div className={styles.emblaSlide} key={video.videoId}>
+              <Videothumb
+                key={video.videoId}
+                videoName={video.name}
+                videoId={video.videoId}
+                videoLink={video.link}
+                hideLabel={hideLabel}
+              />
+            </div>
+          )
+        }
+      </EmblaCarousel>
     </div>
   )
 }
